Use int instead of long in matrix operations to replicate the ORIGINAL behavior

+++ src/poly.h
@@ -66,15 +66,15 @@ struct view_type {
 
 // Global transformation arrays:
 
-long matrix[4][4];          // Master transformation matrix
-long smat[4][4];            // Scaling matrix
-long rmat[4][4];            // Perspective matrix
-long zmat[4][4];            // Z rotation matrix
-long xmat[4][4];            // X rotation matrix
-long ymat[4][4];            // Y rotation matrix
-long tmat[4][4];            // Translation matrix
+int matrix[4][4];          // Master transformation matrix
+int smat[4][4];            // Scaling matrix
+int rmat[4][4];            // Perspective matrix
+int zmat[4][4];            // Z rotation matrix
+int xmat[4][4];            // X rotation matrix
+int ymat[4][4];            // Y rotation matrix
+int tmat[4][4];            // Translation matrix
 
-void matmult(long result[4][4], long mat1[4][4], long mat2[4][4])
+void matmult(int result[4][4], int mat1[4][4], int mat2[4][4])
 {
 	// Multiply matrix MAT1 by matrix MAT2, returning the result in RESULT
 
@@ -85,7 +85,7 @@ void matmult(long result[4][4], long mat1[4][4], long mat2[4][4])
 	}
 }
 
-void matcopy(long dest[4][4], long source[4][4])
+void matcopy(int dest[4][4], int source[4][4])
 {
 	// Copy matrix SOURCE to matrix DEST
 
@@ -108,13 +108,13 @@ void inittrans()
 
 void scale(float xs, float ys, float zs)
 {
-	long mat[4][4];
+	int mat[4][4];
 
 	// Initialize scaling matrix:
-	smat[0][0] = (long)(zs * SHIFT_MULT); smat[0][1] = 0;                       smat[0][2] = 0;                       smat[0][3] = 0;
-	smat[1][0] = 0;                       smat[1][1] = (long)(ys * SHIFT_MULT); smat[1][2] = 0;                       smat[1][3] = 0;
-	smat[2][0] = 0;                       smat[2][1] = 0;                       smat[2][2] = (long)(xs * SHIFT_MULT); smat[2][3] = 0;
-	smat[3][0] = 0;                       smat[3][1] = 0;                       smat[3][2] = 0;                       smat[3][3] = ONE;
+	smat[0][0] = (int)(zs * SHIFT_MULT); smat[0][1] = 0;                      smat[0][2] = 0;                      smat[0][3] = 0;
+	smat[1][0] = 0;                      smat[1][1] = (int)(ys * SHIFT_MULT); smat[1][2] = 0;                      smat[1][3] = 0;
+	smat[2][0] = 0;                      smat[2][1] = 0;                      smat[2][2] = (int)(xs * SHIFT_MULT); smat[2][3] = 0;
+	smat[3][0] = 0;                      smat[3][1] = 0;                      smat[3][2] = 0;                      smat[3][3] = ONE;
 
 	// Concatenate with master matrix:
 	matmult(mat, smat, matrix);
@@ -123,7 +123,7 @@ void scale(float xs, float ys, float zs)
 
 void reflect(int xr, int yr, int zr)
 {
-	long mat[4][4];
+	int mat[4][4];
 
 	rmat[0][0] = xr >> SHIFT; rmat[0][1] = 0;           rmat[0][2] = 0;           rmat[0][3] = 0;
 	rmat[1][0] = 0;           rmat[1][1] = yr >> SHIFT; rmat[1][2] = 0;           rmat[1][3] = 0;
@@ -140,8 +140,8 @@ void rotate(int ax, int ay, int az)
 	// AX radians on the X axis, AY radians on the Y axis and
 	// AZ radians on the Z axis
 
-	long mat1[4][4];
-	long mat2[4][4];
+	int mat1[4][4];
+	int mat2[4][4];
 
 	// Initialize Y rotation matrix:
 	ymat[0][0] = COS(ay); ymat[0][1] = 0;   ymat[0][2] = -SIN(ay); ymat[0][3] = 0;
@@ -160,10 +160,10 @@ void rotate(int ax, int ay, int az)
 	// Concatenate this matrix with master matrix:
 
 	// Initialize Z rotation matrix:
-	zmat[0][0] = COS(az);  zmat[0][1] = SIN(az);  zmat[0][2] = 0;    zmat[0][3] = 0;
-	zmat[1][0] = -SIN(az); zmat[1][1] = COS(az);  zmat[1][2] = 0;    zmat[1][3] = 0;
-	zmat[2][0] = 0;        zmat[2][1] = 0;        zmat[2][2] = ONE;  zmat[2][3] = 0;
-	zmat[3][0] = 0;        zmat[3][1] = 0;        zmat[3][2] = 0;    zmat[3][3] = ONE;
+	zmat[0][0] = COS(az);  zmat[0][1] = SIN(az); zmat[0][2] = 0;   zmat[0][3] = 0;
+	zmat[1][0] = -SIN(az); zmat[1][1] = COS(az); zmat[1][2] = 0;   zmat[1][3] = 0;
+	zmat[2][0] = 0;        zmat[2][1] = 0;       zmat[2][2] = ONE; zmat[2][3] = 0;
+	zmat[3][0] = 0;        zmat[3][1] = 0;       zmat[3][2] = 0;   zmat[3][3] = ONE;
 
 	// Concatenate this matrix with master matrix:
 	matmult(mat1, matrix, ymat);
@@ -177,12 +177,12 @@ void translate(int xt, int yt, int zt)
 	// X distance of XT, a Y distance of YT, and a Z distance of ZT
 	// from the screen origin
 
-	long mat[4][4];
+	int mat[4][4];
 
-	tmat[0][0] = ONE;               tmat[0][1] = 0;                 tmat[0][2] = 0;                     tmat[0][3] = 0;
-	tmat[1][0] = 0;                 tmat[1][1] = ONE;               tmat[1][2] = 0;                     tmat[1][3] = 0;
-	tmat[2][0] = 0;                 tmat[2][1] = 0;                 tmat[2][2] = ONE;                   tmat[2][3] = 0;
-	tmat[3][0] = (long)xt << SHIFT; tmat[3][1] = (long)yt << SHIFT; tmat[3][2] = (long)zt * SHIFT_MULT; tmat[3][3] = ONE;
+	tmat[0][0] = ONE;              tmat[0][1] = 0;                tmat[0][2] = 0;                    tmat[0][3] = 0;
+	tmat[1][0] = 0;                tmat[1][1] = ONE;              tmat[1][2] = 0;                    tmat[1][3] = 0;
+	tmat[2][0] = 0;                tmat[2][1] = 0;                tmat[2][2] = ONE;                  tmat[2][3] = 0;
+	tmat[3][0] = (int)xt << SHIFT; tmat[3][1] = (int)yt << SHIFT; tmat[3][2] = (int)zt * SHIFT_MULT; tmat[3][3] = ONE;
 
 	// Concatenate with master matrix:
 	matmult(mat, matrix, tmat);
@@ -195,9 +195,9 @@ void transform(object_type *object)
 
 	for (int v = 0; v < (*object).number_of_vertices; v++) {
 		vertex_type *vptr = &object->vertex[v];
-		vptr->wx = ((long)vptr->lx * matrix[0][0] + (long)vptr->ly * matrix[1][0] + (long)vptr->lz * matrix[2][0] + matrix[3][0]) >> SHIFT;
-		vptr->wy = ((long)vptr->lx * matrix[0][1] + (long)vptr->ly * matrix[1][1] + (long)vptr->lz * matrix[2][1] + matrix[3][1]) >> SHIFT;
-		vptr->wz = ((long)vptr->lx * matrix[0][2] + (long)vptr->ly * matrix[1][2] + (long)vptr->lz * matrix[2][2] + matrix[3][2]) >> SHIFT;
+		vptr->wx = ((int)vptr->lx * matrix[0][0] + (int)vptr->ly * matrix[1][0] + (int)vptr->lz * matrix[2][0] + matrix[3][0]) >> SHIFT;
+		vptr->wy = ((int)vptr->lx * matrix[0][1] + (int)vptr->ly * matrix[1][1] + (int)vptr->lz * matrix[2][1] + matrix[3][1]) >> SHIFT;
+		vptr->wz = ((int)vptr->lx * matrix[0][2] + (int)vptr->ly * matrix[1][2] + (int)vptr->lz * matrix[2][2] + matrix[3][2]) >> SHIFT;
 	}
 }
 
@@ -207,9 +207,9 @@ void atransform(object_type *object)
 
 	for (int v = 0; v < (*object).number_of_vertices; v++) {
 		vertex_type *vptr = &(*object).vertex[v];
-		vptr->ax = ((long)vptr->wx * matrix[0][0] + (long)vptr->wy * matrix[1][0] + (long)vptr->wz * matrix[2][0] + matrix[3][0]) >> SHIFT;
-		vptr->ay = ((long)vptr->wx * matrix[0][1] + (long)vptr->wy * matrix[1][1] + (long)vptr->wz * matrix[2][1] + matrix[3][1]) >> SHIFT;
-		vptr->az = ((long)vptr->wx * matrix[0][2] + (long)vptr->wy * matrix[1][2] + (long)vptr->wz * matrix[2][2] + matrix[3][2]) >> SHIFT;
+		vptr->ax = ((int)vptr->wx * matrix[0][0] + (int)vptr->wy * matrix[1][0] + (int)vptr->wz * matrix[2][0] + matrix[3][0]) >> SHIFT;
+		vptr->ay = ((int)vptr->wx * matrix[0][1] + (int)vptr->wy * matrix[1][1] + (int)vptr->wz * matrix[2][1] + matrix[3][1]) >> SHIFT;
+		vptr->az = ((int)vptr->wx * matrix[0][2] + (int)vptr->wy * matrix[1][2] + (int)vptr->wz * matrix[2][2] + matrix[3][2]) >> SHIFT;
 	}
 }
